[api] Move TypeFactory into its own type classes
