+++ src/main/kotlin/dev/supergrecko/kllvm/core/types/ArrayType.kt
@@ -23,4 +23,18 @@ public class ArrayType(llvmType: LLVMTypeRef) : Type(llvmType) {
 
         return Type(type)
     }
+
+    public companion object {
+        /**
+         * Create an array type
+         *
+         * Constructs an array of type [ty] with size [size].
+         */
+        @JvmStatic
+        public fun new(type: Type, size: Int): ArrayType {
+            require(size >= 0) { "Cannot make array of negative size" }
+
+            return ArrayType(LLVM.LLVMArrayType(type.llvmType, size))
+        }
+    }
 }
